Удаление папки Lib
