--- VulnerabilityAnalyzerLib/Lib/Models/Report.cs
@@ -1,32 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace Lib.Models
-{
-    public partial class Report
-    {
-        private IReportRepresenter reportRepresenter;
-        private List<ReportNode> reportNodes;
-
- 
-        private Report(List<ReportNode> reportNodes, IReportRepresenter reportRepresenter)
-        {
-            this.reportRepresenter = reportRepresenter;
-            this.reportNodes = reportNodes;
-        }
-
-        async public static Task<Report> CreateReport(
-            List<Vulnerability> vulnerabilities,
-            List<SourceText> sources,
-            IReportRepresenter reportRepresenter)
-        {
-            List<ReportNode> reportNodes = new();
-            // TODO: Finish method
-
-            return new Report(reportNodes, reportRepresenter);
-        }
-
-        public void Dump(string fitePath) => reportRepresenter.Dump(fitePath, reportNodes);
-        public string Dumps() => reportRepresenter.Dumps(reportNodes);
-    }
-}
