--- VulnerabilityAnalyzerLib/Lib/Models/ReportNode.cs
@@ -1,14 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Lib.Models
-{
-    public partial class Report
-    {
-        public class ReportNode
-        {
-            public SourceLine SourceLine { get; set; }
-            public Vulnerability Vulnerability { get; set; }
-            public List<Pattern> PassedPatterns { get; set; }
-        }
-    }
-}
