--- VulnerabilityAnalyzerLib/Lib/Models/SourceLine.cs
@@ -1,15 +1,0 @@
-﻿namespace Lib.Models
-{
-    public class SourceLine
-    {
-        public int LineNumber { get; set; }
-        public string SourcePath { get; set; }
-        public string Line { get; set; }
-
-        public bool ApplyPattern(Pattern pattern)
-        {
-            // TODO
-            return true;
-        }
-    }
-}
