--- VulnerabilityAnalyzerLib/Lib/Models/SourceText.cs
@@ -1,31 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Lib.Models
-{
-    public class SourceText
-    {
-
-        public string SourcePath { get; set; }
-        private List<SourceLine> lines;
-        private int currentLine = 0;
-
-        public SourceText(string sourcePath, List<SourceLine> lines)
-        {
-            this.SourcePath = sourcePath;
-            this.lines = lines;
-        }
-        public SourceLine GetLine()
-        {
-            // TODO: Check there is any line and throw exception otherwise.
-            return lines[currentLine + 1];
-        }
-
-        public bool AllLinesChecked()
-        {
-            // TODO: Finish method
-            return true;
-        }
-
-        public void GoToTextBeginning() => currentLine = 0;
-    }
-}
