--- VulnerabilityAnalyzerLib/Lib/Utils/IReportRepresenter.cs
@@ -1,13 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Lib.Models
-{
-    public partial class Report
-    {
-        public interface IReportRepresenter
-        {
-            public void Dump(string filePath, List<ReportNode> reportNodes);
-            public string Dumps(List<ReportNode> reportNodes);
-        }
-    }
-}
