+++ VulnerabilityAnalyzerLib/Models/Pattern.cs
@@ -1,4 +1,4 @@
-﻿namespace Lib.Models
+﻿namespace Models
 {
     public class Pattern
     {
