+++ VulnerabilityAnalyzerLib/Models/Report.cs
@@ -1,14 +1,14 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
-using Common.Models;
 
-namespace VulnerabilityAnalyzerLib.Models
+namespace Models
 {
     public partial class Report
     {
         private IReportRepresenter reportRepresenter;
         private List<ReportNode> reportNodes;
 
+ 
         private Report(List<ReportNode> reportNodes, IReportRepresenter reportRepresenter)
         {
             this.reportRepresenter = reportRepresenter;
@@ -21,37 +21,12 @@ namespace VulnerabilityAnalyzerLib.Models
             IReportRepresenter reportRepresenter)
         {
             List<ReportNode> reportNodes = new();
-            foreach (Vulnerability vulnerability in vulnerabilities)
-            {
-                foreach (SourceText source in sources)
-                {
-                    reportNodes.AddRange(await FindVulnerabilityInSource(source, vulnerability));
-                }
-            }
+            // TODO: Finish method
+
             return new Report(reportNodes, reportRepresenter);
         }
 
-        async private static Task<List<ReportNode>> FindVulnerabilityInSource(SourceText source, Vulnerability vulnerability)
-        {
-            return await Task.Run(() =>
-            {
-                List<ReportNode> nodes = new();
-                foreach (SourceLine line in source.Lines)
-                {
-                    List<Common.Models.Pattern> patterns = new();
-                    foreach (Common.Models.Pattern pattern in vulnerability.Patterns)
-                    {
-                        if (line.ApplyPattern(pattern))
-                        {
-                            patterns.Add(pattern);
-                        }
-                    }
-                    nodes.Add(new ReportNode() { SourceLine = line, Vulnerability = vulnerability, PassedPatterns = patterns });
-                }
-                return nodes;
-            });
-        }
-        public void Dump(string fitePath) => reportRepresenter.Dump(this, fitePath);
-        public string Dumps() => reportRepresenter.Dumps(this);
+        public void Dump(string fitePath) => reportRepresenter.Dump(fitePath, reportNodes);
+        public string Dumps() => reportRepresenter.Dumps(reportNodes);
     }
 }
