+++ VulnerabilityAnalyzerLib/Models/ReportNode.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections.Generic;
-using Common.Models;
 
-namespace VulnerabilityAnalyzerLib.Models
+namespace Models
 {
     public partial class Report
     {
@@ -9,7 +8,7 @@ namespace VulnerabilityAnalyzerLib.Models
         {
             public SourceLine SourceLine { get; set; }
             public Vulnerability Vulnerability { get; set; }
-            public List<Common.Models.Pattern> PassedPatterns { get; set; }
+            public List<Pattern> PassedPatterns { get; set; }
         }
     }
 }
