+++ VulnerabilityAnalyzerLib/Models/SourceLine.cs
@@ -1,18 +1,15 @@
-﻿using System.Text.RegularExpressions;
-using Common.Models;
-
-namespace VulnerabilityAnalyzerLib.Models
+﻿namespace Models
 {
     public class SourceLine
     {
         public int LineNumber { get; set; }
         public string SourcePath { get; set; }
         public string Line { get; set; }
-        public bool ApplyPattern(Common.Models.Pattern pattern)
+
+        public bool ApplyPattern(Pattern pattern)
         {
-            Regex regex = new Regex(pattern.Mask, RegexOptions.IgnoreCase);
-            MatchCollection matches = regex.Matches(Line);
-            return matches.Count > 0;
+            // TODO
+            return true;
         }
     }
 }
