+++ VulnerabilityAnalyzerLib/Models/Vulnerability.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Lib.Models
+namespace Models
 {
     public class Vulnerability
     {
