+++ VulnerabilityAnalyzerLib/Repository/IVulnerabilityRepository.cs
@@ -1,8 +1,8 @@
-﻿using Lib.Models;
+﻿using Models;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace Lib.Repository
+namespace VulnerabilityAnalyzerLib.Repository
 {
     public interface IVulnerabilityRepository
     {
