+++ VulnerabilityAnalyzerLib/Repository/JsonVulnerabilityRepository.cs
@@ -1,13 +1,11 @@
-﻿using Lib.Models;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text.Json;
 using System.Threading.Tasks;
+using Models;
 
-
-
-namespace Lib.Repository
+namespace VulnerabilityAnalyzerLib.Repository
 {
     public class JsonVulnerabilityRepository : IVulnerabilityRepository
     {
@@ -83,7 +81,7 @@ namespace Lib.Repository
 
         public async Task WriteToFile()
         {
-            var jsonString = JsonSerializer.Serialize<List<Vulnerability>>(_vulnerabilities);
+            var jsonString = JsonSerializer.Serialize(_vulnerabilities);
             await File.WriteAllTextAsync(jsonString, _filePath);
         }
