+++ VulnerabilityAnalyzerLib/Utils/HtmlReportRepresenter.cs
@@ -3,7 +3,7 @@ using System.Data;
 using System.IO;
 using System.Text;
 
-namespace Lib.Models
+namespace Models
 {
     public partial class Report
     {
