+++ VulnerabilityAnalyzerLib/Utils/IReportRepresenter.cs
@@ -1,13 +1,13 @@
 ﻿using System.Collections.Generic;
 
-namespace VulnerabilityAnalyzerLib.Models
+namespace Models
 {
     public partial class Report
     {
         public interface IReportRepresenter
         {
-            public void Dump(Report report,  string filePath);
-            public string Dumps(Report report);
+            public void Dump(string filePath, List<ReportNode> reportNodes);
+            public string Dumps(List<ReportNode> reportNodes);
         }
     }
 }
