+++ VulnerabilityAnalyzerLib/Utils/SourceTextReader.cs
@@ -3,14 +3,15 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using Models;
 
-namespace Lib.Models
+namespace VulnerabilityAnalyzerLib.Utils
 {
     public class SourceTextReader
     {
         public List<SourceText> SourceTexts { get; } = new();
         public List<string> SourceExtensions { get; set; } = new() { "cpp", "c", "h", "hpp" };
-        
+
         async public Task ReadFromFile(string filePath)
         {
             try
@@ -18,14 +19,15 @@ namespace Lib.Models
                 using (StreamReader sr = new StreamReader(filePath))
                 {
                     List<SourceLine> lines = new List<SourceLine>();
-                    while (!sr.EndOfStream) 
+                    while (!sr.EndOfStream)
                     {
                         string line = await sr.ReadLineAsync();
-                        for(int i = 0; i < line.Length; i++) {
-                            lines.Add(new SourceLine() { Line = line,LineNumber = i ,SourcePath = filePath}) ;
-                        }          
+                        for (int i = 0; i < line.Length; i++)
+                        {
+                            lines.Add(new SourceLine() { Line = line, LineNumber = i, SourcePath = filePath });
+                        }
                     }
-                    SourceTexts.Add(new SourceText(filePath,lines));
+                    SourceTexts.Add(new SourceText(filePath, lines));
                 }
             }
             catch (Exception ex)
@@ -39,7 +41,7 @@ namespace Lib.Models
             var files = Directory.GetFiles(dirPath);
             foreach (var file in files)
             {
-                if (SourceExtensions.Contains(Path.GetExtension(file))) 
+                if (SourceExtensions.Contains(Path.GetExtension(file)))
                 {
                     await ReadFromFile(file);
                 }
